AudioFormatDescription: remove docs about encapsulation
Initially there was an idea to use MIME-compatible '+'
syntax for specifying both the encapsulated format and the
format of encapsulation, e.g. 'audio/x-iec61937+audio/ac3'.
However, it seems to create too many caveats when
comparing type values and converting to/from legacy and SDK
types. Thus, it's better to store the encapsulation
information in a separate type 'AudioEncapsulationType'.

Bug: 198000804
Test: m
Change-Id: Ieb50ce6b895e54df7d38e865bc9ae1f14a17789e