Consolidate DisplayContent#assignRelativeLayerForIme
When the window focus or the ime target changes,
DC#assignRelativeLayerForIme will be called to set the
relative layering with IME when the imeTarget is the
non-child application window without in animating state.

Howerver, the animating state checking includes all types
of animation. And that makes when taping an
editor dialog fragment to show IME, then swiping up to
recents will see the IME will behind the dimming because
system thought the ime target is animated by recents animation
so the relative layering didn't be set.

As task switching by recents animation that we would like to keep
the current task window hierarchy and IME relative layer during
switching, modify the animating check with excluding recents
animation.

Bug: 201139555
Bug: 204140533
Test: atest ZOrderingTests#\
       testAssignWindowLayers_ForImeOnAppWithRecentsAnimating
Test: manual as steps:
  1) Launch Settings with editor dialogs fragment
    (E.g. Network & Internet > Private DNS)
  2) Tap focus on Editor to show IME
  3) Tap and hold on navigation bar
  4) Expect IME layer isn't below dimming layer

Merged-In: I3024443c21cfb1c7971f3999640039caea10d0d5
Change-Id: I3024443c21cfb1c7971f3999640039caea10d0d5